requirements.txt
numpy==2.1.3
scipy==1.14.1
scikit-learn==1.5.2
matplotlib==3.9.2
imageio==2.36.0
pillow==11.0.0
tensorflow==2.20.0
opencv-python==4.10.0.84
torch==2.5.1
torchvision==0.20.1
